+++ datalad_container/containers_run.py
@@ -3,6 +3,7 @@
 __docformat__ = 'restructuredtext'
 
 import logging
+import os
 import os.path as op
 
 from datalad.interface.base import Interface
@@ -12,13 +13,19 @@ from datalad.distribution.dataset import datasetmethod
 from datalad.distribution.dataset import require_dataset
 from datalad.interface.utils import eval_results
 
+from datalad.interface.results import get_status_dict
 from datalad.interface.run import Run
+from datalad.interface.run import run_command
 from datalad.interface.run import get_command_pwds
 from datalad.interface.run import normalize_command
 from datalad_container.find_container import find_container
 
 lgr = logging.getLogger("datalad.containers.containers_run")
 
+# Environment variable to be set during execution to possibly
+# inform underlying shim scripts about the original name of
+# the container
+CONTAINER_NAME_ENVVAR = 'DATALAD_CONTAINER_NAME'
 
 _run_params = dict(
     Run._params_,
@@ -46,7 +53,14 @@ class ContainersRun(Interface):
     A command is generated based on the input arguments such that the
     container image itself will be recorded as an input dependency of
     the command execution in the `run` record in the git history.
+
+    During execution the environment variable {name_envvar} is set to the
+    name of the used container.
     """
+
+    _docs_ = dict(
+        name_envvar=CONTAINER_NAME_ENVVAR
+    )
 
     _params_ = _run_params
 
@@ -56,12 +70,16 @@ class ContainersRun(Interface):
     def __call__(cmd, container_name=None, dataset=None,
                  inputs=None, outputs=None, message=None, expand=None,
                  explicit=False, sidecar=None):
+        from mock import patch  # delayed, since takes long (~600ms for yoh)
         pwd, _ = get_command_pwds(dataset)
         ds = require_dataset(dataset, check_installed=True,
                              purpose='run a containerized command execution')
 
         container = find_container(ds, container_name)
         image_path = op.relpath(container["path"], pwd)
+        # container record would contain path to the (sub)dataset containing
+        # it.  If not - take current dataset, as it must be coming from it
+        image_dspath = op.relpath(container.get('parentds', ds.path), pwd)
 
         # sure we could check whether the container image is present,
         # but it might live in a subdataset that isn't even installed yet
@@ -83,25 +101,39 @@ class ContainersRun(Interface):
                     raise ValueError(
                         'cmdexe {!r} is in an old, unsupported format. '
                         'Convert it to a plain string.'.format(callspec))
-
-            cmd = callspec.format(img=image_path, cmd=cmd)
+            try:
+                cmd_kwargs = dict(
+                    img=image_path,
+                    cmd=cmd,
+                    img_dspath=image_dspath,
+                )
+                cmd = callspec.format(**cmd_kwargs)
+            except KeyError as exc:
+                yield get_status_dict(
+                    'run',
+                    ds=ds,
+                    status='error',
+                    message=(
+                        'Unrecognized cmdexec placeholder: %s. '
+                        'See containers-add for information on known ones: %s',
+                        exc,
+                        ", ".join(cmd_kwargs)))
+                return
         else:
             # just prepend and pray
             cmd = container['path'] + ' ' + cmd
 
-        # with amend inputs to also include the container image
-        inputs = (inputs or []) + [image_path]
-
-        # fire!
-        for r in Run.__call__(
-                cmd=cmd,
-                dataset=ds,
-                inputs=inputs,
-                outputs=outputs,
-                message=message,
-                expand=expand,
-                explicit=explicit,
-                sidecar=sidecar,
-                on_failure="ignore",
-                return_type='generator'):
-            yield r
+        with patch.dict('os.environ',
+                        {CONTAINER_NAME_ENVVAR: container['name']}):
+            # fire!
+            for r in run_command(
+                    cmd=cmd,
+                    dataset=dataset or (ds if ds.path == pwd else None),
+                    inputs=inputs,
+                    extra_inputs=[image_path],
+                    outputs=outputs,
+                    message=message,
+                    expand=expand,
+                    explicit=explicit,
+                    sidecar=sidecar):
+                yield r