+++ datalad_container/tests/test_run.py
@@ -1,3 +1,4 @@
+import os
 import os.path as op
 
 from six import text_type
@@ -12,10 +13,16 @@ from datalad.tests.utils import ok_clean_git
 from datalad.tests.utils import assert_in
 from datalad.tests.utils import assert_result_count
 from datalad.tests.utils import assert_raises
+from datalad.tests.utils import ok_file_has_content
 from datalad.tests.utils import with_tempfile
 from datalad.tests.utils import with_tree
 from datalad.tests.utils import skip_if_no_network
-from datalad.utils import on_windows
+from datalad.tests.utils import swallow_outputs
+from datalad.utils import (
+    chpwd,
+    on_windows,
+)
+from datalad.support.network import get_local_file_url
 
 
 testimg_url = 'shub://datalad/datalad-container:testhelper'
@@ -48,6 +55,18 @@ def test_run_mispecified(path):
     assert_in("Container selection impossible", text_type(cm.exception))
 
 
+@with_tree(tree={"i.img": "doesn't matter"})
+def test_run_unknown_cmdexec_placeholder(path):
+    ds = Dataset(path).create(force=True)
+    ds.containers_add("i", image="i.img", call_fmt="{youdontknowme}")
+    assert_result_count(
+        ds.containers_run("doesn't matter", on_failure="ignore"),
+        1,
+        path=ds.path,
+        action="run",
+        status="error")
+
+
 @skip_if_no_network
 @with_tempfile
 @with_tempfile
@@ -70,6 +89,15 @@ def test_container_files(path, super_path):
         updateurl=testimg_url)
     ok_clean_git(path)
 
+    def assert_no_change(res, path):
+        # this command changed nothing
+        #
+        # Avoid specifying the action because it will change from "add" to
+        # "save" in DataLad v0.12.
+        assert_result_count(
+            res, 1, status='notneeded',
+            path=path, type='dataset')
+
     # now we can run stuff in the container
     # and because there is just one, we don't even have to name the container
     res = ds.containers_run(cmd)
@@ -77,9 +105,7 @@ def test_container_files(path, super_path):
     assert_result_count(
         res, 1, action='get', status='notneeded',
         path=op.join(ds.path, 'righthere'), type='file')
-    # this command changed nothing
-    assert_result_count(
-        res, 1, action='add', status='notneeded', path=ds.path, type='dataset')
+    assert_no_change(res, ds.path)
 
     # same thing as we specify the container by its name:
     res = ds.containers_run(cmd,
@@ -88,9 +114,7 @@ def test_container_files(path, super_path):
     assert_result_count(
         res, 1, action='get', status='notneeded',
         path=op.join(ds.path, 'righthere'), type='file')
-    # this command changed nothing
-    assert_result_count(
-        res, 1, action='add', status='notneeded', path=ds.path, type='dataset')
+    assert_no_change(res, ds.path)
 
     # we can also specify the container by its path:
     res = ds.containers_run(cmd,
@@ -99,9 +123,7 @@ def test_container_files(path, super_path):
     assert_result_count(
         res, 1, action='get', status='notneeded',
         path=op.join(ds.path, 'righthere'), type='file')
-    # this command changed nothing
-    assert_result_count(
-        res, 1, action='add', status='notneeded', path=ds.path, type='dataset')
+    assert_no_change(res, ds.path)
 
     # Now, test the same thing, but with this dataset being a subdataset of
     # another one:
@@ -119,7 +141,65 @@ def test_container_files(path, super_path):
     assert_result_count(
         res, 1, action='get', status='ok',
         path=op.join(super_ds.path, 'sub', 'righthere'), type='file')
-    # this command changed nothing
-    assert_result_count(
-        res, 1, action='add', status='notneeded', path=super_ds.path, type='dataset')
+    assert_no_change(res, super_ds.path)
 
+
+@with_tempfile
+@with_tree(tree={'some_container.img': "doesn't matter"})
+def test_custom_call_fmt(path, local_file):
+    ds = Dataset(path).create()
+    subds = ds.create('sub')
+
+    # plug in a proper singularity image
+    subds.containers_add(
+        'mycontainer',
+        url=get_local_file_url(op.join(local_file, 'some_container.img')),
+        image='righthere',
+        call_fmt='echo image={img} cmd={cmd} img_dspath={img_dspath} '
+                 # and environment variable being set/propagated by default
+                 'name=$DATALAD_CONTAINER_NAME'
+    )
+    ds.save()  # record the effect in super-dataset
+
+    # Running should work fine either withing sub or within super
+    with swallow_outputs() as cmo:
+        subds.containers_run('XXX', container_name='mycontainer')
+        assert_in('image=righthere cmd=XXX img_dspath=. name=mycontainer', cmo.out)
+
+    with swallow_outputs() as cmo:
+        ds.containers_run('XXX', container_name='sub/mycontainer')
+        assert_in('image=sub/righthere cmd=XXX img_dspath=sub', cmo.out)
+
+    # Test within subdirectory of the super-dataset
+    subdir = op.join(ds.path, 'subdir')
+    os.mkdir(subdir)
+    with chpwd(subdir):
+        with swallow_outputs() as cmo:
+            containers_run('XXX', container_name='sub/mycontainer')
+            assert_in('image=../sub/righthere cmd=XXX img_dspath=../sub', cmo.out)
+
+
+@skip_if_no_network
+@with_tree(tree={"subdir": {"in": "innards"}})
+def test_run_no_explicit_dataset(path):
+    ds = Dataset(path).create(force=True)
+    ds.add(".")
+    ds.containers_add("deb", url=testimg_url,
+                      call_fmt="singularity exec {img} {cmd}")
+
+    # When no explicit dataset is given, paths are interpreted as relative to
+    # the current working directory.
+
+    # From top-level directory.
+    with chpwd(path):
+        containers_run("cat {inputs[0]} {inputs[0]} >doubled",
+                       inputs=[op.join("subdir", "in")],
+                       outputs=["doubled"])
+        ok_file_has_content(op.join(path, "doubled"), "innardsinnards")
+
+    # From under a subdirectory.
+    subdir = op.join(ds.path, "subdir")
+    with chpwd(subdir):
+        containers_run("cat {inputs[0]} {inputs[0]} >doubled",
+                       inputs=["in"], outputs=["doubled"])
+    ok_file_has_content(op.join(subdir, "doubled"), "innardsinnards")